+++ src/app/(recipe)/layout.jsx
@@ -6,7 +6,7 @@ export default function MyRecipeLayout({ children }) {
   return (
     <div className="mx-auto flex min-h-screen sm:max-w-2xl sm:gap-x-5 sm:px-4">
       <Navigation />
-      <main className="w-full overflow-hidden border-gray-200 pb-5 sm:border-x">
+      <main className="w-full overflow-hidden border-gray-200 pb-20 sm:border-x">
         <Tabs
           tabItems={[
             { label: "作り方", href: "/recipe" },
