--- src/app/(top)/my-page/(_)/_components/index.js
@@ -1,3 +1,0 @@
-export { MenuButton } from "./MenuButton";
-export { ProfileEditLink } from "./ProfileEditLink";
-export { RecipeAddLink } from "./RecipeAddLink";
