+++ src/app/(top)/my-page/(default)/_components/index.js
@@ -1,0 +1,1 @@
+export { MenuButton } from "./MenuButton";
