+++ src/app/(top)/my-page/(default)/page.jsx
@@ -2,9 +2,5 @@ import { VerticalRecipeList } from "@/components/RecipeList";
 
 // マイページ（新着レシピ）
 export default function MyPage() {
-  return (
-    <div className="px-4 pt-5">
-      <VerticalRecipeList />
-    </div>
-  );
+  return <VerticalRecipeList />;
 }
