+++ src/app/(top)/my-page/(default)/popular/page.jsx
@@ -2,9 +2,5 @@ import { VerticalRecipeList } from "@/components/RecipeList";
 
 // マイページ（人気レシピ）
 export default function Popular() {
-  return (
-    <div className="px-4 pt-5">
-      <VerticalRecipeList />
-    </div>
-  );
+  return <VerticalRecipeList />;
 }
