+++ src/app/(top)/new-recipes/page.jsx
@@ -4,7 +4,7 @@ import { VerticalRecipeList } from "@/components/RecipeList";
 // 新着レシピ画面
 export default function NewRecipes() {
   return (
-    <div className="pb-20">
+    <div>
       <div className="px-4 py-3">
         <NewRecipesBar />
       </div>
