+++ src/components/ShoppingList/ShoppingList.jsx
@@ -2,7 +2,7 @@ import { RecipeItem } from "./RecipeItem";
 
 export function ShoppingList() {
   return (
-    <div className="space-y-10 pb-20 pt-5">
+    <div className="space-y-10 pt-5">
       <RecipeItem />
       <RecipeItem />
       <RecipeItem />
